+++ docs/web_modules/layouts/Page/index.js
@@ -1,6 +1,9 @@
 import React, { Component, PropTypes } from "react"
 import Helmet from "react-helmet"
 import invariant from "invariant"
+import { joinUri } from "phenomic"
+
+import EditThisPage from "../../EditThisPage"
 
 import styles from "./index.css"
 
@@ -40,7 +43,10 @@ export default class Page extends Component {
     const meta = [
       { property: "og:type", content: "article" },
       { property: "og:title", content: metaTitle },
-      { property: "og:url", content: __url },
+      {
+        property: "og:url",
+        content: joinUri(process.env.PHENOMIC_USER_URL, __url),
+      },
       { property: "og:description", content: head.description },
       { name: "twitter:card", content: "summary" },
       { name: "twitter:title", content: metaTitle },
@@ -49,16 +55,28 @@ export default class Page extends Component {
       { name: "description", content: head.description },
     ]
 
+    const PageActions = (
+      <div className={ styles.pageActions }>
+        <EditThisPage filename={ __filename } />
+      </div>
+    )
+
     return (
       <div>
         <Helmet
           title={ metaTitle }
           meta={ meta }
         />
-
         {
           head.title &&
           <h1>{ head.title }</h1>
+        }
+        {
+          // to avoid "weird" visual result, we put actions at the top only
+          // if page has a title, other wise (eg: homepage) it can be a little
+          // weird
+          head.title &&
+          PageActions
         }
         {
           head.incomplete &&
@@ -77,6 +95,7 @@ export default class Page extends Component {
           />
         }
         { this.props.children }
+        { PageActions }
       </div>
     )
   }