+++ src/_utils/serialize/index.js
@@ -1,0 +1,5 @@
+// @flow
+
+export default (value: any): string => {
+  return JSON.stringify(value).replace(/\<\/script>/g, "<\\/script>")
+}