+++ src/bin/commands/setup/__tests__/questions.js
@@ -1,15 +1,20 @@
 import test from "ava"
-import questions from "../questions"
+import questions, { validateUrl } from "../questions"
 
 test("should validate name with only letters, numbers and dashes", (t) => {
-  const name = questions.find((q) => q.name === "name")
-  t.true(name.validate("foo-bar-BAZ-1-2-3"))
-  t.true(typeof name.validate("foo .") === "string")
-  t.true(typeof name.validate("BAR 1 .") === "string")
+  const validate = questions.find((q) => q.name === "name").validate
+  t.true(validate("foo-bar-BAZ-1-2-3"))
+  t.true(typeof validate("foo .") === "string")
+  t.true(typeof validate("BAR 1 .") === "string")
 })
 
-test("should validate homepage url", (t) => {
-  const homepage = questions.find((q) => q.name === "homepage")
-  t.true(homepage.validate("http://foo.bar"))
-  t.true(typeof homepage.validate("foo.bar") === "string")
+test("should validate url", (t) => {
+  t.true(validateUrl("http://foo.bar"))
+  t.true(typeof validateUrl("foo.bar") === "string")
+})
+
+test("should validate twitter username", (t) => {
+  const validate = questions.find((q) => q.name === "twitter").validate
+  t.true(validate("Phenomic_app"))
+  t.true(typeof validate("Phenomic app") === "string")
 })