+++ src/bin/commands/setup/__tests__/stub/spawn-prompt.js
@@ -1,9 +1,14 @@
 require("babel-register")
 const inquirer = require("inquirer")
-const questions = require("../../questions").default
 // For testing purpose only
 // Spawn a new process with inquirer auto acive
 // Receive "Done" message indicate that test is passed
+const questions = {
+  type: "input",
+  name: "name",
+  message: "In 1 word describe phenomic",
+}
+
 inquirer
   .prompt(questions)
   .then(() => console.log("Done"))