+++ src/bin/commands/setup/template.js
@@ -3,6 +3,8 @@ const template = {
   "private": true,
   "name": undefined,
   "homepage": undefined,
+  "twitter": undefined,
+  "repository": undefined,
   "scripts": {
     "lint:js": "eslint --ignore-path .gitignore --fix .",
     "lint:css": "stylelint \"web_modules/**/*.css\"",
@@ -15,10 +17,17 @@ const template = {
   "phenomic": {/* placeholder */},
   "babel": {
     "presets": [
-      "react",
-      "es2015",
-      "stage-1",
+      "babel-preset-react",
+      "babel-preset-es2015",
+      "babel-preset-stage-1",
     ],
+    "env": {
+      "production": {
+        "presets": [
+          "babel-preset-react-optimize",
+        ],
+      },
+    },
   },
   "eslintConfig": {
     "parser": "babel-eslint",